fix: Address code review findings for autoscaler and burst resilience
- Fix service name mismatch in CronJob autoscaler URL (use hardcoded
  incidentfox-agent to match Service definition)
- Change ttl_hours type annotation from int to float on 3 methods
  (create_sandbox, create_sandbox_claim, create_sandbox_from_pool)
- Add error handling for SANDBOX_TTL_MINUTES env var parse on follow-up path
- Add security documentation to /metrics/sandbox-demand endpoint

+++ sre-agent/sandbox_manager.py
@@ -406,7 +406,7 @@ static_resources:
         thread_id: str,
         tenant_id: str = "local",
         team_id: str = "local",
-        ttl_hours: int = 2,
+        ttl_hours: float = 2,
         jwt_token: Optional[str] = None,
         team_token: Optional[str] = None,
     ) -> SandboxInfo:
@@ -1147,7 +1147,7 @@ static_resources:
         thread_id: str,
         tenant_id: str,
         team_id: str,
-        ttl_hours: int = 2,
+        ttl_hours: float = 2,
     ) -> tuple[str, str]:
         """
         Create a SandboxClaim to bind to a warm pool pod.
@@ -1408,7 +1408,7 @@ static_resources:
         thread_id: str,
         tenant_id: str = "local",
         team_id: str = "local",
-        ttl_hours: int = 2,
+        ttl_hours: float = 2,
         jwt_token: Optional[str] = None,
         team_token: Optional[str] = None,
     ) -> SandboxInfo:
